+++ src/pygpsclient/gnss_status.py
@@ -44,3 +44,4 @@ class GNSSStatus:
         self.diff_station = "N/A"  # DGPS station id
         self.gsv_data = []  # list of satellite tuples (gnssId, svid, elev, azim, cno)
         self.spectrum_data = []  # list of spectrum data (spec, spn, res, ctr, pga)
+        self.sysmon_data = {}  # dict of system monitor data (cpu and memory load, etc.)